Added ViewBox wrapping for all the visual elements in EntitiesPresenter to resize accordingly

+++ EntitiesPresenter/Models/EntityModel.cs
@@ -9,7 +9,7 @@ namespace EntitiesPresenter.Models
     public class EntityModel
     {
         public string? Name { get; set; }
-        public int X { get; set; }
-        public int Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
     }
 }
